requirements.txt
msgpack==1.0.4
py3rijndael==0.3.3
mitmproxy==10.2.0
orjson==3.9.10
python-dotenv==1.0.0
https://github.com/squaresmile/cppdael/archive/v0.0.2.6.tar.gz
